Added dark contact variable and renamed icon to Green -> GreenTick
