+++ src/components/UI/Form/InlineInput/InlineInput.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect, createRef } from 'react';
 import { TextField } from '@material-ui/core';
-import { ReactComponent as Icon } from '../../../../assets/images/icons/Green.svg';
+import { ReactComponent as Icon } from '../../../../assets/images/icons/GreenTick.svg';
 
 import styles from './InlineInput.module.css';
